# Performance gains on `no-custom-classname` and `no-contradicting-classname`