--- a/lib/rules/no-contradicting-classname.js
+++ b/lib/rules/no-contradicting-classname.js
@@ -61,6 +61,9 @@ module.exports = {
     // Helpers
     //----------------------------------------------------------------------
 
+    // Init assets before sorting
+    const groups = groupUtil.getGroups(defaultGroups, mergedConfig);
+
     /**
      * Parse the classnames and report found conflicts
      * @param {Array} classNames
@@ -69,7 +72,6 @@ module.exports = {
     const parseForContradictingClassNames = (classNames, node) => {
       classNames = attrUtil.sanitizeClassnames(classNames);
       // Init assets before sorting
-      const groups = groupUtil.getGroups(defaultGroups, mergedConfig);
       const sorted = groupUtil.initGroupSlots(groups);
 
       // Move each classname inside its dedicated group