--- a/lib/rules/no-custom-classname.js
+++ b/lib/rules/no-custom-classname.js
@@ -74,6 +74,10 @@ module.exports = {
     // Helpers
     //----------------------------------------------------------------------
 
+    // Init assets before sorting
+    const groups = groupUtil.getGroups(defaultGroups, mergedConfig);
+    const classnamesFromFiles = getClassnamesFromCSS(cssFiles);
+
     /**
      * Parse the classnames and report found conflicts
      * @param {Array} classNames
@@ -81,9 +85,6 @@ module.exports = {
      */
     const parseForCustomClassNames = (classNames, node) => {
       classNames = attrUtil.sanitizeClassnames(classNames);
-      // Init assets before sorting
-      const groups = groupUtil.getGroups(defaultGroups, mergedConfig);
-      const classnamesFromFiles = getClassnamesFromCSS(cssFiles);
 
       classNames.forEach((className) => {
         const idx = groupUtil.getGroupIndex(className, groups, mergedConfig.separator);