@francoismassart 2021-07-06
Hi @larrifax,
I did publish your changes in a beta release: [1.13.5-beta](https://github.com/francoismassart/eslint-plugin-tailwindcss/tree/1.13.5-beta).

I'll try it tomorrow on my current project.

You can install it via
`npm i --save eslint-plugin-tailwindcss@beta`